Add a GitHub workflow to build and deploy a given package
This workflow most notably does not try anything like `please.sh
upgrade` of the git-for-windows/build-extra repository, which tries to
figure out the latest component version and tries to upgrade to that.

Instead, this workflow builds a given package, for the specified
architecture, at the very version committed to the specified branch. The
workflow needs to be triggered manually.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>