build-and-deploy: allow debugging with tmate
The `tmate` program is a combination of `tmux` and `ssh`, allowing
users to connect interactively e.g. to GitHub's build agents.

With this change, we introduce a new build-time variable
`debug_with_ssh_key`. If set, and if the build fails, the workflow will
start a `tmate` instance accepting only connections using the specified
SSH key, to enable the user to debug interactively what went wrong.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>