gnutls(i686): make sure to use the `msys-` prefix
Since MSYS2 no longer maintains an i686 flavor, Git for Windows' i686
SDK is getting increasingly out of sync. One symptom is that the GNUTLS
library was built using the `cyg` prefix instead of the `msys-` one.
This broke `dirmngr` by essentially removing the DLL it relies on.
Again.

Let's work around this by copying the version of `libtool` that we know
works, and be done with it. For now.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>