bash: stop providing `.pdb` files
There seems to be a problem with code-signing the `sh.exe` provided by
the Git for Windows project, and I suspect that it has something to do
with running `cv2pdb` on the executable.

And since nobody, ever, made good use out of these `bash.pdb` files
anyway, let's just stop building them.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>