msys2-runtime: allow patches to work with CR/LF
A recent patch from the `cygwin-3_3_branch` touches `.appveyor.xml`,
which has CR/LF endings.

However, all `.patch` files are checked in with the `text` attribute
set, which means that the CR/LF line endings in that patch are converted
to LF line endings, and the patch actually no longer applies.

Let's force `git am` to ignore whitespace changes for that reason.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>