perl: prepare for v5.36
This not only rebases the patches, but also fixes a all test failures.
This required dropping that `sed` call from `PKGBUILD` that messed up
all trailing whitespace (including the whitespace required by the tests,
e.g. where TAB-delimited lines ended in an empty value, i.e. in a
horizontal tab).

At the same time, `./patchprov` was run, which pointed out that there
was still a "patch" to a file in the `Pod-Parser` module even while that
module was no longer included even in Perl v5.32.x (and hence that patch
actually created the no-longer-needed file).

In the process, a bug was fixed in `patchprov` where it downcased all
Perl package names.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>