openssh: prepare for v9.1
The first patch did not apply.

While at it, prepare for a potential future change where the MSYS2
runtime starts respecting the `CYGWIN` variable if the `MSYS` variable
is unset (https://github.com/msys2/msys2-runtime/issues/70), by not only
forcing `MSYS` to be empty in the `bsd-cygwin_util, but also `CYGWIN`.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>