bash: avoid using the ancient i686 libreadline
In the i686 variation of Git for Windows' SDK, most dependencies have
not seen any updates because we use MSYS2's packages for the most part
and MSYS2 stopped supporting i686 MSYS packages altogether.

However, Git for Windows needs to support i686 builds for a little while
longer and therefore must make do with what is available.

A young enough libreadline ain't it. It's not available. Except in-tree,
there is a libreadline that looks modern enough to use. So let's do
that, but only for i686 builds of Bash.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>