msys2-runtime (update-patches): avoid unnecessary patches
The patches we carry in the `msys2-runtime/` subdirectory are supposed
to reflect the commits in msys2/msys2-runtime of the respective branch.

But not all of those patches make sense. For example, we do not need any
GitHub workflow definitions in these here patches, as those definitions
would not be used while building the MSYS2 runtime anyway.

So let's reduce the clutter and drop those patches explicitly.

While at it, fix some "funny" indentation.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>