msys2-runtime: avoid sharing a cygheap between different MSYS2 runtime versions
When calling, say, a Cygwin Bash (or even a regular MSYS2 Bash, or even
a MinGit Bash of a different Git for Windows version), it is possible
that the MSYS2 runtime of the called Bash is incompatible with the MSYS2
runtime of the caller. When that happens, all kinds of chaos can ensue.

One possible symptom is that the MSYS2 runtime aborts with the error
message "cygheap base mismatch detected".

Another possible symptom is that the MSYS2 runtime (or the Cygwin
runtime) exits silently, with exit code 127 (which is misleading as that
exit code is reserved to imply that an executable could not be found).

With this patch, we try harder to prevent such a thing from happening.

This is a companion to https://github.com/git-for-windows/msys2-runtime/pull/48

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>