msys2-runtime: backport fix for hang-up of `less.exe`
This backports a fix from Cygwin that avoids a hang in `less.exe` after
quitting the pager when the window was resized during the paging.

This fixes https://github.com/git-for-windows/git/issues/3438 (and
https://github.com/git-for-windows/git/issues/4060 and
https://github.com/git-for-windows/git/issues/3683).

The commit looks a bit more active than just integrating
https://github.com/git-for-windows/msys2-runtime/pull/47 because the
script to update the patches now picks up only relevant patches and
skips the useless patch signature, too.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>