msys2-runtime: report the current msys2-runtime commit
With this change, we prepare the `./configure` call to make use of the
new feature of the MSYS2 runtime to report the current msys2-runtime
commit, and also to use it when determining whether another MSYS2
runtime instance's cygheap is compatible with the current MSYS2 runtime.

This requires https://github.com/git-for-windows/msys2-runtime/pull/49's
patches to be brought into MSYS2-packages before it has any effect.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>