libgpg-error: stop pretending to build for Cygwin
The MSYS2 project recently adjusted their build infrastructure to allow
pretending to build for Cygwin but nevertheless produce MSYS binaries.

This works well for x86_64. But for i686, MSYS2 support was stopped a
long time ago, and therefore it does not work.

Git for Windows does need to support i686, still, though. So let's
configure the libgpg-error package for the same CHOST as we used to.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>